Move summary page to its own folder
